pr feedback: extract common bulk fn to common util
